backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-auth==2.23.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
